+++ src/components/layout/index.js
@@ -2,5 +2,6 @@ import AppContainer from './AppContainer/AppContainer';
 import Container from './Container/Container';
 import Header from './Header/Header';
 import ScrollToTop from './ScrollToTop/ScrollToTop';
+import Separator from './Separator/Separator';
 
-export { AppContainer, Container, Header, ScrollToTop };
+export { AppContainer, Container, Header, ScrollToTop, Separator };
